Refactor: Add callbacks, move Beamer initiliser
