+++ src/utils/external/index.ts
@@ -111,32 +111,3 @@ export const getBlockExplorerLink = ({
 
 export const getNetworkRelaseLink = (version: ColonyVersion) =>
   `${NETWORK_RELEASES}/${releaseMap[version]}`;
-
-// Add support for Beamer (getbeamer.com) in app feature announcements
-export const getBeamerId = process.env.BEAMER_PRODUCT_ID || false;
-
-export const getBeamerInitialize = (id: string, url?: string, args = {}) => {
-  if (!window) {
-    return;
-  }
-
-  if (!id) {
-    throw Error('Must provide Beamer "id".');
-  }
-
-  /* eslint-disable camelcase */
-  (window as any).beamer_config = {
-    product_id: id,
-    ...args,
-  };
-  /* eslint-enable camelcase */
-
-  const beamerURL = url || 'https://app.getbeamer.com/js/beamer-embed.js';
-
-  const head = document.getElementsByTagName('head')[0];
-  const script = document.createElement('script');
-
-  script.defer = true;
-  script.src = beamerURL;
-  head.appendChild(script);
-};
